app/requirements.txt
PyPDF2
pinecone-client
sentence-transformers
pinecone-client